+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/Tabs.1.doc.mdx
@@ -3,11 +3,11 @@ import { Badge } from 'terra-tabs/package.json?dev-site-package';
 import TabsPropsTable from 'terra-tabs/src/Tabs?dev-site-props-table';
 import PanePropsTable from 'terra-tabs/src/TabPane?dev-site-props-table';
 
-import TabsTemplate from '../example/TabsTemplate?dev-site-example';
-import TabsTemplateExpanded from '../example/TabsTemplateExpanded?dev-site-example';
-import TabsWithFilledContent from '../example/TabsWithFilledContent?dev-site-example';
-import IconOnlyTabs from '../example/IconOnlyTabs?dev-site-example';
-import ResponsiveTabsVariant from '../example/ResponsiveTabsVariant?dev-site-example';
+import TabsTemplate from './example/TabsTemplate?dev-site-example';
+import TabsTemplateExpanded from './example/TabsTemplateExpanded?dev-site-example';
+import TabsWithFilledContent from './example/TabsWithFilledContent?dev-site-example';
+import IconOnlyTabs from './example/IconOnlyTabs?dev-site-example';
+import ResponsiveTabsVariant from './example/ResponsiveTabsVariant?dev-site-example';
 
 <Badge />
