+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/ResponsiveTabsVariant.jsx
