+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/ResponsiveTabsVariantWrapper.module.scss
