+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/TabsWithFilledContent.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Tabs from 'terra-tabs';
-import TabContent from 'terra-tabs/lib/terra-dev-site/doc/example/TabContentTemplate';
+import TabContent from './TabContentTemplate';
 import styles from './common/TabExample.module.scss';
 
 const cx = classNames.bind(styles);
