+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/common/TabExample.module.scss
