+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/IconAndTextTabPane.test.jsx
@@ -1,17 +1,19 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import TabPane from '../../../../TabPane';
+import IconSearch from 'terra-icon/lib/icon/IconSearch';
+import TabPane from 'terra-tabs';
 import styles from './common/TabPane.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const TextOnlyTabPane = () => (
+const IconAndTextTabPane = () => (
   <div role="tablist" className={cx('content-wrapper')}>
     <TabPane
-      label="Text only tab pane"
-      id="textOnly"
+      label="Icon and Text"
+      id="iconAndText"
+      icon={<IconSearch />}
     />
   </div>
 );
 
-export default TextOnlyTabPane;
+export default IconAndTextTabPane;
