+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/IconOnlyTabPane.test.jsx
@@ -1,19 +1,20 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import IconSearch from 'terra-icon/lib/icon/IconSearch';
-import TabPane from '../../../../TabPane';
+import TabPane from 'terra-tabs';
 import styles from './common/TabPane.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const IconAndTextTabPane = () => (
+const IconOnlyTabPane = () => (
   <div role="tablist" className={cx('content-wrapper')}>
     <TabPane
-      label="Icon and Text"
-      id="iconAndText"
+      label="Icon Only"
+      id="iconOnly"
       icon={<IconSearch />}
+      isIconOnly
     />
   </div>
 );
 
-export default IconAndTextTabPane;
+export default IconOnlyTabPane;
