+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/LongTextTabPane.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import TabPane from '../../../../TabPane';
+import TabPane from 'terra-tabs';
 import styles from './common/TabPane.test.module.scss';
 
 const cx = classNames.bind(styles);
