+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/TextOnlyTabPane.test.jsx
@@ -1,20 +1,17 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import IconSearch from 'terra-icon/lib/icon/IconSearch';
-import TabPane from '../../../../TabPane';
+import TabPane from 'terra-tabs';
 import styles from './common/TabPane.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const IconOnlyTabPane = () => (
+const TextOnlyTabPane = () => (
   <div role="tablist" className={cx('content-wrapper')}>
     <TabPane
-      label="Icon Only"
-      id="iconOnly"
-      icon={<IconSearch />}
-      isIconOnly
+      label="Text only tab pane"
+      id="textOnly"
     />
   </div>
 );
 
-export default IconOnlyTabPane;
+export default TextOnlyTabPane;
