+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/common/TabPane.test.module.scss
