+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsResponsiveToParent.test.jsx
