+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsResponsiveToWindow.test.jsx
