+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsTemplate.jsx
