+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/UncollapsedTabs.test.jsx
